@@ -1,6 +1,6 @@
 module github.com/Goscord/goscord
 
-go 1.14
+go 1.18
 
 require (
 	github.com/asaskevich/EventBus v0.0.0-20200907212545-49d423059eef
